ObjMethod: Always pass context.Context for API method calls
All API server calls make passing context.Context optional. This means
that most if not all API requests do not utilise the context available
to them. This means that it makes it a lot hard to know when a request
should be cancelled or discarded.

This will become really useful when migrating a way from mongo and to
another database store. Additionally we can start to hang more things
from the context allow them to pass through the stack.

This is a breaking change, by forcing context.Context to become
non-optional and forcing juju to handle context correctly.