Allow method removals
The following allows the removal of a method. This helps discard methods
which we don't want to export for any particular reason. As part of the
discarding process, we also add it to the DiscardedMethods.

This is compatible with what is already exists, so we don't need to bump
the revision/branch.

Added tests that was missed when I extracted this first time around.